# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FlightHub API Documentation",
        "description": "API pour la recherche et la réservation de vols",
        "contact": {
            "name": "FlightHub Support",
            "email": "support@flighthub.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement local"
        }
    ],
    "paths": {
...
        "/trips": {
            "get": {
                "tags": [
                    "Trips"
                ],
                "summary": "Récupérer la liste des voyages",
                "description": "Retourne la liste des voyages avec tri et pagination",
                "operationId": "getTrips",
                "parameters": [
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Champ de tri",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "created_at"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Ordre de tri",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des voyages récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PaginatedResponse"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Trips retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Trips"
                ],
                "summary": "Créer un nouveau voyage",
                "description": "Crée un nouveau voyage avec les vols sélectionnés",
                "operationId": "createTrip",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "trip_type": {
                                        "description": "Type de voyage",
                                        "type": "string",
                                        "enum": [
                                            "one-way",
                                            "round-trip",
                                            "multi-city"
                                        ],
                                        "example": "one-way"
                                    },
                                    "flights": {
                                        "description": "IDs des vols",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2
                                        ]
                                    },
                                    "departure_date": {
                                        "description": "Date de départ",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-15"
                                    },
                                    "return_date": {
                                        "description": "Date de retour (optionnel)",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-20"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Voyage créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Trip"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Trip created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données de validation invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/trips/{trip}": {
            "get": {
                "tags": [
                    "Trips"
                ],
                "summary": "Récupérer un voyage spécifique",
                "description": "Retourne les détails d'un voyage par son ID",
                "operationId": "getTrip",
                "parameters": [
                    {
                        "name": "trip",
                        "in": "path",
                        "description": "ID du voyage",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Voyage récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Trip"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Trip retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/trips/validate": {
            "post": {
                "tags": [
                    "Trips"
                ],
                "summary": "Valider un voyage",
                "description": "Valide les contraintes d'un voyage avant création",
                "operationId": "validateTrip",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "trip_type": {
                                        "description": "Type de voyage",
                                        "type": "string",
                                        "enum": [
                                            "one-way",
                                            "round-trip",
                                            "multi-city"
                                        ],
                                        "example": "one-way"
                                    },
                                    "flights": {
                                        "description": "IDs des vols",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2
                                        ]
                                    },
                                    "departure_date": {
                                        "description": "Date de départ",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-15"
                                    },
                                    "return_date": {
                                        "description": "Date de retour (optionnel)",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-20"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Voyage validé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "valid": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Trip is valid"
                                        },
                                        "total_price": {
                                            "description": "Prix total du voyage",
                                            "type": "number",
                                            "format": "float",
                                            "example": 599.98
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Voyage invalide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "valid": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Trip validation failed"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "PaginatedResponse": {
                "title": "Réponse Paginée",
                "description": "Structure de réponse paginée",
                "properties": {
                    "current_page": {
                        "description": "Page actuelle",
                        "type": "integer",
                        "example": 1
                    },
                    "data": {
                        "description": "Données de la page",
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "first_page_url": {
                        "description": "URL de la première page",
                        "type": "string",
                        "example": "http://localhost:8000/api/flights?page=1"
                    },
                    "from": {
                        "description": "Premier élément de la page",
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "description": "Dernière page",
                        "type": "integer",
                        "example": 5
                    },
                    "last_page_url": {
                        "description": "URL de la dernière page",
                        "type": "string",
                        "example": "http://localhost:8000/api/flights?page=5"
                    },
                    "next_page_url": {
                        "description": "URL de la page suivante",
                        "type": "string",
                        "example": "http://localhost:8000/api/flights?page=2"
                    },
                    "path": {
                        "description": "URL de base",
                        "type": "string",
                        "example": "http://localhost:8000/api/flights"
                    },
                    "per_page": {
                        "description": "Éléments par page",
                        "type": "integer",
                        "example": 10
                    },
                    "prev_page_url": {
                        "description": "URL de la page précédente",
                        "type": "string",
                        "example": null
                    },
                    "to": {
                        "description": "Dernier élément de la page",
                        "type": "integer",
                        "example": 10
                    },
                    "total": {
                        "description": "Nombre total d'éléments",
                        "type": "integer",
                        "example": 50
                    }
                },
                "type": "object"
            },
...
            "Flight": {
                "title": "Vol",
                "description": "Modèle représentant un vol",
                "properties": {
                    "id": {
                        "description": "ID unique du vol",
                        "type": "integer",
                        "example": 1
                    },
                    "flight_number": {
                        "description": "Numéro de vol",
                        "type": "string",
                        "example": "AC123"
                    },
                    "airline_id": {
                        "description": "ID de la compagnie aérienne",
                        "type": "integer",
                        "example": 1
                    },
                    "departure_airport_id": {
                        "description": "ID de l'aéroport de départ",
                        "type": "integer",
                        "example": 1
                    },
                    "arrival_airport_id": {
                        "description": "ID de l'aéroport d'arrivée",
                        "type": "integer",
                        "example": 2
                    },
                    "price": {
                        "description": "Prix du vol",
                        "type": "number",
                        "format": "float",
                        "example": 299.99
                    },
                    "departure_time": {
                        "description": "Heure de départ",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:30:00Z"
                    },
                    "arrival_time": {
                        "description": "Heure d'arrivée",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T12:45:00Z"
                    },
                    "airline": {
                        "$ref": "#/components/schemas/Airline"
                    },
                    "departure_airport": {
                        "$ref": "#/components/schemas/Airport"
                    },
                    "arrival_airport": {
                        "$ref": "#/components/schemas/Airport"
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de mise à jour",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Trip": {
                "title": "Voyage",
                "description": "Modèle représentant un voyage",
                "properties": {
                    "id": {
                        "description": "ID unique du voyage",
                        "type": "integer",
                        "example": 1
                    },
                    "trip_type": {
                        "description": "Type de voyage",
                        "type": "string",
                        "enum": [
                            "one-way",
                            "round-trip",
                            "multi-city"
                        ],
                        "example": "one-way"
                    },
                    "total_price": {
                        "description": "Prix total du voyage",
                        "type": "number",
                        "format": "float",
                        "example": 599.98
                    },
                    "departure_date": {
                        "description": "Date de départ",
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-15"
                    },
                    "return_date": {
                        "description": "Date de retour (optionnel)",
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-20"
                    },
                    "flights": {
                        "description": "Vols du voyage",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Flight"
                        }
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de mise à jour",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}